Another Refactor
Cleaned up Arceus a bunch more.
