+++ lib/commands/jump.dart
@@ -6,6 +6,7 @@ import 'package:args/command_runner.dart';
 
 import 'package:cli_spin/cli_spin.dart';
 import 'package:interact/interact.dart';
+import 'package:arceus/extensions.dart';
 
 class JumpCommand extends Command with GetRest {
   @override
@@ -13,7 +14,14 @@ class JumpCommand extends Command with GetRest {
 
   @override
   String get description => """
-jump to a different star in a constellation. Will check for changes in the tracked folder.
+Jump to a different star in a constellation. Will check for changes in the tracked folder.
+Actions with X will default to 1, and you can chain all of these commands with a comma.
+
+Available Actions:
+  - recent: Jump to most recent star.
+  - root: Jump to root star.
+  - forward X: Jump forward by X stars.
+  - back X: Jump backward by X stars.
 """;
 
   @override
@@ -26,14 +34,14 @@ jump to a different star in a constellation. Will check for changes in the track
 
   @override
   Future<void> run() async {
+    final constName = findOption("const");
     final commands = getRest("Enter the star to jump to.");
-    CliSpin spinner = CliSpin(
-            text: "Attempting jump to $commands...", spinner: CliSpinners.moon)
-        .start();
-    final kit = SKit("${Arceus.constFolderPath}/${argResults?["const"]}.skit");
+    CliSpin spinner =
+        CliSpin(text: "Checking for changes...", spinner: CliSpinners.moon)
+            .start();
+    final kit = SKit("${Arceus.constFolderPath}/$constName.skit");
     final constellation = await kit.getConstellation();
-    final archive = await constellation!.getCurrentStar().archive;
-    if (await archive!.checkForChanges(constellation.path)) {
+    if (await constellation!.checkForChanges()) {
       spinner.warn("There are changes in the tracked folder. ");
       final confirm = Confirm(
               prompt:
@@ -43,13 +51,12 @@ jump to a different star in a constellation. Will check for changes in the track
       if (!confirm) {
         return;
       }
-      spinner = CliSpin(
-              text: "Overwriting changes in the tracked folder...",
-              spinner: CliSpinners.moon)
-          .start();
     }
-    constellation.getStarAt(commands).makeCurrent();
-    await archive.extract(constellation.path);
+    spinner = CliSpin(
+            text: "Attempting jump to $commands...", spinner: CliSpinners.moon)
+        .start();
+    final star = constellation.getStarAt(commands);
+    await star.makeCurrent();
     await kit.save();
     spinner.success("Jumped to ${constellation.getCurrentStar().name}!");
   }
